🎨 style: format of the code

+++ domain/src/main/kotlin/cn/soybean/domain/system/entity/SystemMenuEntity.kt
@@ -90,21 +90,27 @@ open class SystemMenuEntity(
      */
     @Column(name = "built_in", nullable = false)
     val builtIn: Boolean = false,
-) : BaseEntity(), PanacheEntityBase {
+) : BaseEntity(),
+    PanacheEntityBase {
     companion object : PanacheCompanion<SystemMenuEntity> {
-        fun listMenuIdByRoleId(roleId: String, userId: String, tenantId: String): Uni<List<String>> = when {
-            isSuperUser(userId) && isSuperRole(roleId) -> listAll().map { menus -> menus.map { it.id } }
+        fun listMenuIdByRoleId(
+            roleId: String,
+            userId: String,
+            tenantId: String,
+        ): Uni<List<String>> =
+            when {
+                isSuperUser(userId) && isSuperRole(roleId) -> listAll().map { menus -> menus.map { it.id } }
 
-            else ->
-                list(
-                    """
+                else ->
+                    list(
+                        """
                        SELECT m FROM SystemMenuEntity m
                        LEFT JOIN SystemRoleMenuEntity rm ON rm.menuId = m.id
                        WHERE rm.roleId = ?1 AND rm.tenantId = ?2
                 """,
-                    roleId,
-                    tenantId,
-                ).map { menus -> menus.map { it.id } }
-        }
+                        roleId,
+                        tenantId,
+                    ).map { menus -> menus.map { it.id } }
+            }
     }
 }
